Add (very unsatisfactory) workaround for IDOX cookie issue
This opens a new tab the first time the person tries.
AJAX requests to fetch the page via AJAX will fail, because of CORS.
Frame loading also fails, because of X-Frame-Options.
Pop-under fails, because of standard browser denial.
Auto-detecting external site onload fails, because of CORS.
Only option therefore is a visible page load, to an expected 5 seconds, then auto-close.
It is hoped the user will then retry, finding it works.